pyproject.toml
[tool.poetry]
name = "llm-search"
version = "0.1.0"
description = "LLM-powered search application with Gmail integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.1.0"
google-api-python-client = "^2.100.0"
sentence-transformers = "^2.2.2"
qdrant-client = "^1.7.0"
tqdm = "^4.66.0"
email_reply_parser = "^0.5.12"
PyPDF2 = "^3.0.1"
python-docx = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 110

[tool.flake8]
max-line-length = 110
extend-ignore = ["E203", "E701", "W503", "E501"]
